# Conflicts:
#	src/main/java/com/example/back_end/model/Customers.java

+++ src/main/java/com/example/back_end/controller/ContractRestController.java
@@ -1,9 +1,10 @@
 package com.example.back_end.controller;
 
+import com.example.back_end.dto.ContractDto;
 import com.example.back_end.dto.CreateContractDto;
 import com.example.back_end.model.Contracts;
 import com.example.back_end.projections.ContractSearchDTO;
-import com.example.back_end.projections.IContractProjection;
+import com.example.back_end.projections.ITransactionHistoryProjection;
 import com.example.back_end.service.IContractService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,46 +18,33 @@ import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
-/**
- * Created by: DinhHD
- * Date created: 13/07/2023
- * Function: do about pawn interface, customer selection interface
- *
-// * @param Contracts
- * @return createContracts()
- */
-
 @RequestMapping("/api/employee/contract")
 @RestController
-
 @CrossOrigin("*")
 public class ContractRestController {
     @Autowired
     private IContractService iContractService;
 
-    /**
-     * Created by: ThienNT
-     * Date created: 13/07/2023
-     * Function: get page transaction history from Database
-     * <p>
-     *
-     * @param page
-     * @return ResponseEntity<Page < IContractProjection>>
-     */
+//    /**
+//     * Created by: ThienNT
+//     * Date created: 13/07/2023
+//     * Function: get page transaction history from Database
+//     * <p>
+//     *
+//     * @param
+//     * @return ResponseEntity<Page < IContractProjection>>
+//     */
 
-    @GetMapping("")
-    public ResponseEntity<Page<IContractProjection>> getALlTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
-                                                                              @RequestParam(name = "limit", defaultValue = "5") Integer limit) {
-        Page<IContractProjection> contractProjectionPage = this.iContractService.findAllTransactionHistory(page, limit);
-        int totalPage = contractProjectionPage.getTotalPages();
-        if (page >= totalPage) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-        if (contractProjectionPage.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        }
-        return new ResponseEntity<>(contractProjectionPage, HttpStatus.OK);
-    }
+//    @GetMapping("")
+//    public ResponseEntity<Page<ITransactionHistoryProjection>> getAllTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
+//                                                                                        @RequestParam(name = "limit", defaultValue = "5") Integer limit) {
+//        Page<ITransactionHistoryProjection> contractProjectionPage = this.iContractService.findAllTransactionHistory(page, limit);
+//        int totalPage = contractProjectionPage.getTotalPages();
+//        if (page >= totalPage || page < 0) {
+//            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+//        }
+//        return new ResponseEntity<>(contractProjectionPage, HttpStatus.OK);
+//    }
 
     /**
      * Created by: ThienNT
@@ -69,7 +57,11 @@ public class ContractRestController {
      */
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Boolean> deleteTransactionHistoryById(@PathVariable("id") Integer id) {
+    public ResponseEntity<Boolean> deleteTransactionHistoryById(@PathVariable("id") String id) {
+        Optional<Contracts> contractDTO = iContractService.findTransactionHistoryById(id);
+        if (!contractDTO.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         Boolean check = iContractService.deleteTransactionHistoryById(id);
         return new ResponseEntity<>(check, HttpStatus.OK);
     }
@@ -85,9 +77,9 @@ public class ContractRestController {
      */
 
     @GetMapping("/detail/{id}")
-    public ResponseEntity<Contracts> getTransactionHistoryDetail(@PathVariable("id") Integer id) {
+    public ResponseEntity<Contracts> showTransactionHistoryDetail(@PathVariable("id") String id) {
         Optional<Contracts> contractDTO = iContractService.findTransactionHistoryById(id);
-        return contractDTO.map(iContractProjection -> new ResponseEntity<>(iContractProjection, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.BAD_REQUEST));
+        return contractDTO.map(contracts -> new ResponseEntity<>(contracts, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.BAD_REQUEST));
     }
 
     /**
@@ -96,22 +88,19 @@ public class ContractRestController {
      * Function: search transaction history from Database
      * <p>
      *
-     * @param contractSearchDTO
+     * @param
      * @return ResponseEntity<IContractProjection>
      * @requestbody contractSearchDTO
      */
 
-    @GetMapping("/search")
-    public ResponseEntity<Page<IContractProjection>> searchTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
-                                                                              @RequestParam(name = "limit", defaultValue = "5") Integer limit,
-                                                                              @RequestBody ContractSearchDTO contractSearchDTO) {
-        Page<IContractProjection> contractProjectionsPage = iContractService.searchTransactionHistory(page, limit, contractSearchDTO);
+    @PostMapping("/search-transaction-history")
+    public ResponseEntity<Page<ITransactionHistoryProjection>> showListAndSearchTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
+                                                                                                   @RequestParam(name = "limit", defaultValue = "5") Integer limit,
+                                                                                                   @RequestBody ContractSearchDTO contractSearchDTO) {
+        Page<ITransactionHistoryProjection> contractProjectionsPage = iContractService.showListAndSearchTransactionHistory(page, limit, contractSearchDTO);
         int totalPage = contractProjectionsPage.getTotalPages();
-        if (page >= totalPage) {
+        if (page > totalPage || page<0) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-        if (contractProjectionsPage.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
         return new ResponseEntity<>(contractProjectionsPage, HttpStatus.OK);
     }
@@ -123,13 +112,49 @@ public class ContractRestController {
     }
 
     @PostMapping("/createContract")
-    public ResponseEntity<?> createContracts(@RequestBody @Valid CreateContractDto contractDto, BindingResult bindingResult){
-        if (bindingResult.hasErrors()){
+    public ResponseEntity<HttpStatus> createContracts(@RequestBody @Valid CreateContractDto contractDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-        Contracts contracts=new Contracts();
-        BeanUtils.copyProperties(contractDto,contracts);
-            iContractService.createContract(contracts);
-            return new ResponseEntity<>(HttpStatus.CREATED);
+        Contracts contracts = new Contracts();
+        BeanUtils.copyProperties(contractDto, contracts);
+        iContractService.createContract(contracts);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+
+    /**
+     * Create by : TriPD
+     * Date created : 13/07/2023
+     * Function : findContractById(),updateContract()
+     */
+
+    @GetMapping("/findContractById/{id}")
+    public ResponseEntity<ContractDto> getContractById(@PathVariable Long id) {
+        Contracts contract = this.iContractService.findContractById(id);
+        if (contract == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        ContractDto contractDto = new ContractDto();
+        BeanUtils.copyProperties(contract, contractDto);
+        return new ResponseEntity<>(contractDto, HttpStatus.OK);
+    }
+
+
+    @PatchMapping("/update")
+    public ResponseEntity<ContractDto> updateContract(@RequestBody ContractDto contractDto) {
+        if (contractDto == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        iContractService.saveContract(contractDto);
+        return new ResponseEntity<>(contractDto, HttpStatus.OK);
+    }
+
+    @GetMapping("/top10")
+    public ResponseEntity<List<Contracts>> top10NewContract() {
+        List<Contracts> contracts = this.iContractService.showTop10NewContract();
+        if (contracts.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(contracts, HttpStatus.OK);
     }
 }
